+++ plugins/in_stdin/in_stdin.c
@@ -123,7 +123,7 @@ int in_stdin_collect(struct flb_config *config, void *in_context)
     return 0;
 }
 
-void *in_stdin_flush(void *in_context, int *size)
+void *in_stdin_flush(void *in_context, size_t *size)
 {
     char *buf;
     msgpack_sbuffer *sbuf;
